requirements_joint_and_metric.txt
setuptools
librosa
addict
appdirs== 1.4.4
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blobfile==2.0.1
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
contourpy==1.2.1
cryptography==39.0.1
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.0
einops-exts==0.0.4
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.9.0
flit-core==3.6.0
fsspec==2024.2.0
ftfy==6.1.1
gdown==5.2.0
gitdb==4.0.10
gitpython==3.1.31
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
intel-openmp==2024.1.2
ipython==8.12.3
jedi==0.19.1
jinja2==3.1.3
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.5
lazy-loader==0.1
llvmlite==0.41.0
lpips==0.1.4
lxml==4.9.2
markupsafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
mkl==2024.1.0
mkl-service==2.4.0
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.4
moviepy
mpmath==1.3.0
msgpack==1.0.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.0
numpy==1.25.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openmim==0.3.9
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==9.4.0
pipreqs==0.5.0
platformdirs==3.5.1
pooch==1.7.0
proglog==0.1.10
prompt-toolkit==3.0.47
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.8
pycparser==2.21
pycryptodomex==3.17
pygments==2.18.0
pyopenssl==23.0.0
pyparsing==3.1.2
pysocks==1.7.1
python-dateutil==2.8.2
pytorch-ignite==0.4.11
pytz==2022.7.1
pyyaml==6.0
pyzmq==26.0.3
referencing==0.35.1
regex==2022.10.31
requests==2.28.1
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.2.1
scipy==1.10.1
sentry-sdk==1.16.0
setproctitle==1.3.2
setuptools==40.8.0
shapely==2.0.4
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.4 
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
termcolor==2.2.0
terminaltables==3.1.10
threadpoolctl==3.1.0
timm==1.0.7
tinycss2==1.3.0
tornado==6.2
tqdm==4.65.0
traitlets==5.14.3
triton==2.1.0
typing-extensions==4.4.0
tzdata==2024.1
urllib3==1.26.14
visdom==0.2.4
wandb==0.13.11
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
wheel==0.38.4
xmltodict==0.13.0
yapf==0.40.2
yarg==0.1.9
zipp==3.15.0
